Changed statistics format for Clone and Extraction passes.
Clone pass statistics report input independent instructions and basic block coverage before running cloning pass and
after the run. It also reports number of clonned instructions and functions.

Extraction pass reports input dependent instructions and basic blocks coverage, which are inside an input dependent
function, both before the extraction and after it. It also reports number of extracted instructions, number or mediate
instructions and the list of extracted functions.

Fixed an issue in the function deciding if basic block is input dependent for the given set of input dependent arguments
Fixed the same issue for the function deciding input dependency of a loop for given set of input dep arguments.