Implemented issue #27
Introduced new key in a function entry in library function configuration file. This entry contains indices of library function
arguments that are of function type and library function turns them to callback, that can be called from input dependent
locations. New entry looks like "callback_arguments": [1], which means 1st argument of given library function will
become input dependent callback.
Added corresponding parser to LibraryInfoFromConfigFile to parse callback_arguments.
Added corresponding fields to DependencyAnalysisResult to collect map from llvm value to potential callback function.
When a function is assigned to a value, this pair is added to the map. When a library function is called with a value
that has entry in the map, corresponding functions are marked input dependent.