Implemented issue #34
Added -transparent-caching flag. If ran with this flag input dependency information is cached using metadata.
Module flag metadata is added to indicate if caching is done. For each function a metadata is added corresponding to
input dependency of the function. Same for basic blocks and instructions. In case if block is input dependent or
unreachable, no metadata is added for instruction of that block, as those are input dependent or unreachable
respectively.