Implemented issue #22
If input dependency pass is ran with -transparent-caching flag and module has input dependency
information cached in the bitcode, instead of running input dependency analysis a cached information is parsed.
If -transparent-caching flag is existent but there is no cached information in the module, input dependency analysis
will run and its information will be cached in the module.

Updated dot printer pass to work with new cached input dependency information.