Fixed an issue in function extraction. Extraction was resulting in malformed function - issue #38
An input independent load instruction from one block was used as a call argument in another - input dependent block
which was extracted as a function. Extracted function was malformed as it has a instruction used but not defined in its
scope(load). Modified implementation to pass load (or any other instruction) as an argument to the extracted function if
it has use there.

Modified merging of snippets to use indices of snippets instead of iterators