issue #42 - Fixed some causes of crashes for tinyxml
Start snippet with a basic block with invoke terminating instruction
Fixed issue in instruction snippet contains_instruction function. Before checking based on instruction index, check if
instruction in the block of snippet or not.
When removing dummy_block after erasing all blocks of a snippet make sure to iterate through remaining uses of dummy
block and drop its reference, given that those uses are in the snippet themselves and will be removed shortly.