Implemented issue #60
Modified rules of extracting block for exception CFG. If any of the exception handling blocks is not included in a
snippet, do not perform extraction. This is a workaround for issue #61
Do not extract snippet if it ends with landing pad block and the block is not contained in the snippet, that is is not
going to be extracted.
Modified remapping of extracted block instruction to be done after remapping start snippet, as blocks coming after can
use values from start snippet and would need to use already remapped ones.
Modified used value collection to check if instruction/value is in a parent block.
Modified check if snippet can be extracted to check if snippet contains instruction operands too. That is if an
instruction from a snippet has an operand which is not inside the snippet, snippet can not be extracted.
Set personality to extracted function is original function has one.